services/python-risk-engine/src/app.py
from fastapi import FastAPI, HTTPException, Depends
from fastapi.middleware.cors import CORSMiddleware
from typing import Dict, Any, List
import logging
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(
    title="Python Risk Engine",
    description="Service for calculating and managing investment risk metrics",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Health check endpoint
@app.get("/health")
async def health_check():
    """Health check endpoint for load balancers and monitoring."""
    return {
        "status": "healthy",
        "service": "risk-engine",
        "version": "1.0.0"
    }

# Example risk calculation endpoint
@app.post("/calculate-risk")
async def calculate_risk(portfolio_data: Dict[str, Any]):
    """Calculate risk metrics for a given portfolio."""
    try:
        # TODO: Implement actual risk calculation logic
        return {
            "status": "success",
            "risk_metrics": {
                "value_at_risk": 0.0,
                "expected_shortfall": 0.0,
                "sharpe_ratio": 0.0,
                "beta": 0.0
            }
        }
    except Exception as e:
        logger.error(f"Error calculating risk: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

if __name__ == "__main__":
    import uvicorn
    uvicorn.run("app:app", host="0.0.0.0", port=8000, reload=True)
